+++ anvil_consortium_manager/tests/test_forms.py
@@ -19,6 +19,12 @@ class BillingProjectImportFormTest(TestCase):
         form = self.form_class(data=form_data)
         self.assertTrue(form.is_valid())
 
+    def test_form_valid_note(self):
+        """Form is valid with the note field."""
+        form_data = {"name": "foo", "note": "test note"}
+        form = self.form_class(data=form_data)
+        self.assertTrue(form.is_valid())
+
     def test_invalid_missing_name(self):
         """Form is invalid when missing name."""
         form_data = {}
@@ -66,6 +72,16 @@ class AccountImportFormTest(TestCase):
         form_data = {
             "email": "test_email@example.com",
             "is_service_account": True,
+        }
+        form = self.form_class(data=form_data)
+        self.assertTrue(form.is_valid())
+
+    def test_valid_with_note(self):
+        """Form is valid with necessary input."""
+        form_data = {
+            "email": "test_email@example.com",
+            "is_service_account": True,
+            "note": "test note",
         }
         form = self.form_class(data=form_data)
         self.assertTrue(form.is_valid())
@@ -163,6 +179,15 @@ class ManagedGroupCreateFormTest(TestCase):
         form = self.form_class(data=form_data)
         self.assertTrue(form.is_valid())
 
+    def test_valid_with_note(self):
+        """Form is valid with necessary input and note is specified."""
+        form_data = {
+            "name": "test-group-name",
+            "note": "test note",
+        }
+        form = self.form_class(data=form_data)
+        self.assertTrue(form.is_valid())
+
     def test_invalid_missing_name(self):
         """Form is invalid when missing name."""
         form_data = {}
@@ -225,6 +250,17 @@ class WorkspaceCreateFormTest(TestCase):
         form = self.form_class(data=form_data)
         self.assertTrue(form.is_valid())
 
+    def test_valid_with_note(self):
+        """Form is valid with necessary input and note is specified."""
+        billing_project = factories.BillingProjectFactory.create()
+        form_data = {
+            "billing_project": billing_project,
+            "name": "test-workspace",
+            "note": "test note",
+        }
+        form = self.form_class(data=form_data)
+        self.assertTrue(form.is_valid())
+
     def test_invalid_missing_billing_project(self):
         """Form is invalid when missing billing_project_name."""
         form_data = {"name": "test-workspace"}
@@ -283,6 +319,18 @@ class WorkspaceImportFormTest(TestCase):
         """Form is valid with necessary input."""
         form_data = {
             "workspace": "test-billing-project/test-workspace",
+        }
+        workspace_choices = [
+            ("test-billing-project/test-workspace", 1),
+        ]
+        form = self.form_class(workspace_choices=workspace_choices, data=form_data)
+        self.assertTrue(form.is_valid())
+
+    def test_valid_with_note(self):
+        """Form is valid with necessary input and note is specified."""
+        form_data = {
+            "workspace": "test-billing-project/test-workspace",
+            "note": "test note",
         }
         workspace_choices = [
             ("test-billing-project/test-workspace", 1),